requirements.txt
# Core Dependencies
fastapi>=0.120.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# RAG & Vector Store
langchain>=1.1.0
langchain-community>=0.4.0
langchain-openai>=1.0.0
langchain-google-genai>=3.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# Document Processing
pypdf>=4.0.0
tiktoken>=0.5.0

# Speech to Text
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0
soundfile>=0.12.1
pydub>=0.25.1

# Multilingual Support
langdetect>=1.0.9

# Utilities
numpy>=1.24.0
requests>=2.31.0

# Logging & Monitoring
loguru>=0.7.2
